Add johncena package under java
