+++ build.gradle
@@ -29,7 +29,7 @@ test {
 }
 
 application {
-    mainClass.set("JohnCena")
+    mainClass.set("johncena.JohnCena")
 }
 
 shadowJar {
