--- src/main/java/commands/Command.java
@@ -1,8 +1,0 @@
-package commands;
-
-import exceptions.CenaException;
-
-public interface Command {
-
-    void execute() throws CenaException;
-}
