--- src/main/java/commands/HelpCommand.java
@@ -1,22 +1,0 @@
-package commands;
-
-public class HelpCommand implements Command {
-
-    @Override
-    public void execute() {
-        System.out.println("____________________________________________________________");
-        System.out.println("Here are the commands you can use:");
-        System.out.println("  bye - Exits the program");
-        System.out.println("  list - Lists all tasks");
-        System.out.println("  mark [task number] - Marks a task as done");
-        System.out.println("  unmark [task number] - Marks a task as not done");
-        System.out.println("  delete [task number] - Deletes a task");
-        System.out.println("  todo [description] - Adds a todo task");
-        System.out.println("  deadline [description] /by [due date] - Adds a deadline task");
-        System.out.println("  event [description] /from [start date] /to [end date] - Adds an event task");
-        System.out.println("  on [date] - Lists all tasks on a specific date");
-        System.out.println("  hello - Displays the welcome message");
-        System.out.println("  help - Displays the list of commands");
-        System.out.println("____________________________________________________________");
-    }
-}
