+++ src/main/java/johncena/art/Logo.java
@@ -1,0 +1,27 @@
+
+package johncena.art;
+
+public class Logo {
+
+    private static final String logo =
+            "          JJJJJJJJJJJ              hhhhhhh                                           CCCCCCCCCCCCC                                                        \n"
+                    + "          J:::::::::J              h:::::h                                        CCC::::::::::::C                                                        \n"
+                    + "          J:::::::::J              h:::::h                                      CC:::::::::::::::C                                                        \n"
+                    + "          JJ:::::::JJ              h:::::h                                     C:::::CCCCCCCC::::C                                                        \n"
+                    + "            J:::::J   ooooooooooo   h::::h hhhhh      nnnn  nnnnnnnn          C:::::C       CCCCCC    eeeeeeeeeeee    nnnn  nnnnnnnn      aaaaaaaaaaaaa   \n"
+                    + "            J:::::J oo:::::::::::oo h::::hh:::::hhh   n:::nn::::::::nn       C:::::C                ee::::::::::::ee  n:::nn::::::::nn    a::::::::::::a  \n"
+                    + "            J:::::Jo:::::::::::::::oh::::::::::::::hh n::::::::::::::nn      C:::::C               e::::::eeeee:::::een::::::::::::::nn   aaaaaaaaa:::::a \n"
+                    + "            J:::::jo:::::ooooo:::::oh:::::::hhh::::::hnn:::::::::::::::n     C:::::C              e::::::e     e:::::enn:::::::::::::::n           a::::a \n"
+                    + "            J:::::Jo::::o     o::::oh::::::h   h::::::h n:::::nnnn:::::n     C:::::C              e:::::::eeeee::::::e  n:::::nnnn:::::n    aaaaaaa:::::a \n"
+                    + "JJJJJJJ     J:::::Jo::::o     o::::oh:::::h     h:::::h n::::n    n::::n     C:::::C              e:::::::::::::::::e   n::::n    n::::n  aa::::::::::::a \n"
+                    + "J:::::J     J:::::Jo::::o     o::::oh:::::h     h:::::h n::::n    n::::n     C:::::C              e::::::eeeeeeeeeee    n::::n    n::::n a::::aaaa::::::a \n"
+                    + "J::::::J   J::::::Jo::::o     o::::oh:::::h     h:::::h n::::n    n::::n      C:::::C       CCCCCCe:::::::e             n::::n    n::::na::::a    a:::::a \n"
+                    + "J:::::::JJJ:::::::Jo:::::ooooo:::::oh:::::h     h:::::h n::::n    n::::n       C:::::CCCCCCCC::::Ce::::::::e            n::::n    n::::na::::a    a:::::a \n"
+                    + " JJ:::::::::::::JJ o:::::::::::::::oh:::::h     h:::::h n::::n    n::::n        CC:::::::::::::::C e::::::::eeeeeeee    n::::n    n::::na:::::aaaa::::::a \n"
+                    + "   JJ:::::::::JJ    oo:::::::::::oo h:::::h     h:::::h n::::n    n::::n          CCC::::::::::::C  ee:::::::::::::e    n::::n    n::::n a::::::::::aa:::a\n"
+                    + "     JJJJJJJJJ        ooooooooooo   hhhhhhh     hhhhhhh nnnnnn    nnnnnn             CCCCCCCCCCCCC    eeeeeeeeeeeeee    nnnnnn    nnnnnn  aaaaaaaaaa  aaaa\n";
+
+    public static String getLogo() {
+        return logo;
+    }
+}
