+++ src/main/java/johncena/commands/Command.java
@@ -1,0 +1,8 @@
+package johncena.commands;
+
+import johncena.exceptions.CenaException;
+
+public interface Command {
+
+    void execute() throws CenaException;
+}
