+++ src/main/java/johncena/commands/MarkCommand.java
@@ -1,0 +1,30 @@
+package johncena.commands;
+
+import johncena.exceptions.CenaInvalidTaskIndexException;
+import johncena.storage.Storage;
+import johncena.tasks.Task;
+
+import java.util.ArrayList;
+
+public class MarkCommand implements Command {
+    private ArrayList<Task> tasks;
+    private int taskIndex;
+
+    public MarkCommand(ArrayList<Task> tasks, int taskIndex) {
+        this.tasks = tasks;
+        this.taskIndex = taskIndex;
+    }
+
+    @Override
+    public void execute() throws CenaInvalidTaskIndexException {
+        if (taskIndex < 0 || taskIndex >= tasks.size()) {
+            throw new CenaInvalidTaskIndexException("The task index is invalid.");
+        }
+        tasks.get(taskIndex).markAsDone();
+        Storage.saveTasks(tasks);
+        System.out.println("____________________________________________________________");
+        System.out.println(" Nice! I've marked this task as done:");
+        System.out.println("   " + tasks.get(taskIndex));
+        System.out.println("____________________________________________________________");
+    }
+}
